【RestAP sqlite化】マージしたらコンパイルエラーになったので修正

+++ app/src/main/java/com/v3/basis/blas/ui/fixture/fixture_view/FixtureCellModel.kt
@@ -4,7 +4,7 @@ import androidx.databinding.ObservableBoolean
 import androidx.databinding.ObservableField
 
 data class FixtureCellModel(
-    val fixture_id: Int,
+    val fixture_id: Long,
     val title: String = "",
     val detail: String = "") {
